Add build/include/wasmtime
Following the update to Wasmtime's new C API in #81, the header files are now not
only in build/local, but also in build/local/wasmtime.

This commit adds the directory "build/include/wasmtime" as a valid empty go
package, like it was done in #45 for the other build directories.

It also addapts `ci/local.sh` to:
- copy the additional header files in build/include/wasmtime
- stop this script from removing the "empty.go" files.